Set Multi-tier Referral system
