--- contracts/TrustMLM/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "hardhat-project",
-  "devDependencies": {
-    "@nomicfoundation/hardhat-toolbox-viem": "^3.0.0",
-    "@openzeppelin/contracts": "^5.0.2",
-    "@openzeppelin/contracts-upgradeable": "^5.0.2",
-    "@openzeppelin/hardhat-upgrades": "^3.1.0",
-    "erc721a": "^4.3.0",
-    "hardhat": "^2.22.3"
-  }
-}
